# zend_execute: Improve type checking performance when passing a `Closure` to a `callable` parameter