@iluuu1994 2025-04-08
```
Summary
  sapi/cli/php-old callable-micro-opt/closure.php ran
    1.03 ± 0.02 times faster than sapi/cli/php callable-micro-opt/callable.php
    1.04 ± 0.02 times faster than sapi/cli/php callable-micro-opt/closure.php
    1.29 ± 0.02 times faster than sapi/cli/php-old callable-micro-opt/callable.php

Summary
  sapi/cli/php-old callable-micro-opt/native.php ran
    1.01 ± 0.02 times faster than sapi/cli/php callable-micro-opt/native.php

Summary
  sapi/cli/php-old callable-micro-opt/scalar.php ran
    1.02 ± 0.01 times faster than sapi/cli/php callable-micro-opt/scalar.php

Summary
  sapi/cli/php-old callable-micro-opt/union.php ran
    1.02 ± 0.01 times faster than sapi/cli/php callable-micro-opt/union.php

Summary
  sapi/cli/php-old callable-micro-opt/union_obj.php ran
    1.04 ± 0.01 times faster than sapi/cli/php callable-micro-opt/union_obj.php
```

With the exception of callable.php, which was significantly faster on the new version, most other benchmark seem to have slightly regressed. This result is consistent, I tried each benchmark multiple times.